feat: ai poc
The promotion script now saves a html file with summary of
changes and a list of linked commits and the github workflow
will save this as an artifact. That way, when you run the
promotion workflow, you can open the html page afterwards
and copy and paste to gmail to send to users.

Signed-off-by: Martin Malina <mmalina@redhat.com>